refactor: Book/Keywords API 목데이터와 엔드포인트 분리 (#59)
