+++ src/features/keywords/index.ts
@@ -1,3 +1,4 @@
 export * from './hooks'
 export * from './keywords.api'
+export * from './keywords.endpoints'
 export * from './keywords.types'
