+++ src/features/keywords/keywords.endpoints.ts
@@ -1,0 +1,6 @@
+import { API_PATHS } from '@/api'
+
+export const KEYWORDS_ENDPOINTS = {
+  // 키워드 목록 조회 (GET /api/keywords)
+  LIST: API_PATHS.KEYWORDS,
+} as const
