+++ src/features/keywords/keywords.mock.ts
@@ -1,0 +1,731 @@
+/**
+ * @file keywords.mock.ts
+ * @description 키워드 API 목데이터
+ */
+
+import type { GetKeywordsResponse } from './keywords.types'
+
+// ============================================================
+// Mock Data
+// ============================================================
+
+const mockKeywordsResponse: GetKeywordsResponse = {
+  keywords: [
+    // ============================================================
+    // 책 키워드 카테고리 (BOOK, level 1)
+    // ============================================================
+    {
+      id: 1,
+      name: '인간관계',
+      type: 'BOOK',
+      parentId: null,
+      parentName: null,
+      level: 1,
+      sortOrder: 1,
+      isSelectable: false,
+    },
+    {
+      id: 2,
+      name: '개인',
+      type: 'BOOK',
+      parentId: null,
+      parentName: null,
+      level: 1,
+      sortOrder: 2,
+      isSelectable: false,
+    },
+    {
+      id: 3,
+      name: '삶과 죽음',
+      type: 'BOOK',
+      parentId: null,
+      parentName: null,
+      level: 1,
+      sortOrder: 3,
+      isSelectable: false,
+    },
+    {
+      id: 4,
+      name: '사회',
+      type: 'BOOK',
+      parentId: null,
+      parentName: null,
+      level: 1,
+      sortOrder: 4,
+      isSelectable: false,
+    },
+    {
+      id: 5,
+      name: '기타',
+      type: 'BOOK',
+      parentId: null,
+      parentName: null,
+      level: 1,
+      sortOrder: 5,
+      isSelectable: false,
+    },
+
+    // ============================================================
+    // 인간관계 키워드 (level 2, parentId: 1)
+    // ============================================================
+    {
+      id: 11,
+      name: '사랑',
+      type: 'BOOK',
+      parentId: 1,
+      parentName: '인간관계',
+      level: 2,
+      sortOrder: 1,
+      isSelectable: true,
+    },
+    {
+      id: 12,
+      name: '관계',
+      type: 'BOOK',
+      parentId: 1,
+      parentName: '인간관계',
+      level: 2,
+      sortOrder: 2,
+      isSelectable: true,
+    },
+    {
+      id: 13,
+      name: '가족',
+      type: 'BOOK',
+      parentId: 1,
+      parentName: '인간관계',
+      level: 2,
+      sortOrder: 3,
+      isSelectable: true,
+    },
+    {
+      id: 14,
+      name: '우정',
+      type: 'BOOK',
+      parentId: 1,
+      parentName: '인간관계',
+      level: 2,
+      sortOrder: 4,
+      isSelectable: true,
+    },
+    {
+      id: 15,
+      name: '이별',
+      type: 'BOOK',
+      parentId: 1,
+      parentName: '인간관계',
+      level: 2,
+      sortOrder: 5,
+      isSelectable: true,
+    },
+
+    // ============================================================
+    // 개인 키워드 (level 2, parentId: 2)
+    // ============================================================
+    {
+      id: 21,
+      name: '성장',
+      type: 'BOOK',
+      parentId: 2,
+      parentName: '개인',
+      level: 2,
+      sortOrder: 1,
+      isSelectable: true,
+    },
+    {
+      id: 22,
+      name: '자아',
+      type: 'BOOK',
+      parentId: 2,
+      parentName: '개인',
+      level: 2,
+      sortOrder: 2,
+      isSelectable: true,
+    },
+    {
+      id: 23,
+      name: '고독',
+      type: 'BOOK',
+      parentId: 2,
+      parentName: '개인',
+      level: 2,
+      sortOrder: 3,
+      isSelectable: true,
+    },
+    {
+      id: 24,
+      name: '선택',
+      type: 'BOOK',
+      parentId: 2,
+      parentName: '개인',
+      level: 2,
+      sortOrder: 4,
+      isSelectable: true,
+    },
+    {
+      id: 25,
+      name: '자유',
+      type: 'BOOK',
+      parentId: 2,
+      parentName: '개인',
+      level: 2,
+      sortOrder: 5,
+      isSelectable: true,
+    },
+
+    // ============================================================
+    // 삶과 죽음 키워드 (level 2, parentId: 3)
+    // ============================================================
+    {
+      id: 31,
+      name: '삶',
+      type: 'BOOK',
+      parentId: 3,
+      parentName: '삶과 죽음',
+      level: 2,
+      sortOrder: 1,
+      isSelectable: true,
+    },
+    {
+      id: 32,
+      name: '죽음',
+      type: 'BOOK',
+      parentId: 3,
+      parentName: '삶과 죽음',
+      level: 2,
+      sortOrder: 2,
+      isSelectable: true,
+    },
+    {
+      id: 33,
+      name: '상실',
+      type: 'BOOK',
+      parentId: 3,
+      parentName: '삶과 죽음',
+      level: 2,
+      sortOrder: 3,
+      isSelectable: true,
+    },
+    {
+      id: 34,
+      name: '치유',
+      type: 'BOOK',
+      parentId: 3,
+      parentName: '삶과 죽음',
+      level: 2,
+      sortOrder: 4,
+      isSelectable: true,
+    },
+    {
+      id: 35,
+      name: '기억',
+      type: 'BOOK',
+      parentId: 3,
+      parentName: '삶과 죽음',
+      level: 2,
+      sortOrder: 5,
+      isSelectable: true,
+    },
+
+    // ============================================================
+    // 사회 키워드 (level 2, parentId: 4)
+    // ============================================================
+    {
+      id: 41,
+      name: '사회',
+      type: 'BOOK',
+      parentId: 4,
+      parentName: '사회',
+      level: 2,
+      sortOrder: 1,
+      isSelectable: true,
+    },
+    {
+      id: 42,
+      name: '현실',
+      type: 'BOOK',
+      parentId: 4,
+      parentName: '사회',
+      level: 2,
+      sortOrder: 2,
+      isSelectable: true,
+    },
+    {
+      id: 43,
+      name: '역사',
+      type: 'BOOK',
+      parentId: 4,
+      parentName: '사회',
+      level: 2,
+      sortOrder: 3,
+      isSelectable: true,
+    },
+    {
+      id: 44,
+      name: '노동',
+      type: 'BOOK',
+      parentId: 4,
+      parentName: '사회',
+      level: 2,
+      sortOrder: 4,
+      isSelectable: true,
+    },
+    {
+      id: 45,
+      name: '여성',
+      type: 'BOOK',
+      parentId: 4,
+      parentName: '사회',
+      level: 2,
+      sortOrder: 5,
+      isSelectable: true,
+    },
+    {
+      id: 46,
+      name: '윤리',
+      type: 'BOOK',
+      parentId: 4,
+      parentName: '사회',
+      level: 2,
+      sortOrder: 6,
+      isSelectable: true,
+    },
+
+    // ============================================================
+    // 기타 키워드 (level 2, parentId: 5)
+    // ============================================================
+    {
+      id: 51,
+      name: '청춘',
+      type: 'BOOK',
+      parentId: 5,
+      parentName: '기타',
+      level: 2,
+      sortOrder: 1,
+      isSelectable: true,
+    },
+    {
+      id: 52,
+      name: '모험',
+      type: 'BOOK',
+      parentId: 5,
+      parentName: '기타',
+      level: 2,
+      sortOrder: 2,
+      isSelectable: true,
+    },
+    {
+      id: 53,
+      name: '판타지',
+      type: 'BOOK',
+      parentId: 5,
+      parentName: '기타',
+      level: 2,
+      sortOrder: 3,
+      isSelectable: true,
+    },
+    {
+      id: 54,
+      name: '추리',
+      type: 'BOOK',
+      parentId: 5,
+      parentName: '기타',
+      level: 2,
+      sortOrder: 4,
+      isSelectable: true,
+    },
+
+    // ============================================================
+    // 감상 키워드 카테고리 (IMPRESSION, level 1)
+    // ============================================================
+    {
+      id: 101,
+      name: '긍정',
+      type: 'IMPRESSION',
+      parentId: null,
+      parentName: null,
+      level: 1,
+      sortOrder: 1,
+      isSelectable: false,
+    },
+    {
+      id: 102,
+      name: '감상',
+      type: 'IMPRESSION',
+      parentId: null,
+      parentName: null,
+      level: 1,
+      sortOrder: 2,
+      isSelectable: false,
+    },
+    {
+      id: 103,
+      name: '부정',
+      type: 'IMPRESSION',
+      parentId: null,
+      parentName: null,
+      level: 1,
+      sortOrder: 3,
+      isSelectable: false,
+    },
+
+    // ============================================================
+    // 긍정 키워드 (level 2, parentId: 101)
+    // ============================================================
+    {
+      id: 111,
+      name: '즐거운',
+      type: 'IMPRESSION',
+      parentId: 101,
+      parentName: '긍정',
+      level: 2,
+      sortOrder: 1,
+      isSelectable: true,
+    },
+    {
+      id: 112,
+      name: '감동적인',
+      type: 'IMPRESSION',
+      parentId: 101,
+      parentName: '긍정',
+      level: 2,
+      sortOrder: 2,
+      isSelectable: true,
+    },
+    {
+      id: 113,
+      name: '위로받은',
+      type: 'IMPRESSION',
+      parentId: 101,
+      parentName: '긍정',
+      level: 2,
+      sortOrder: 3,
+      isSelectable: true,
+    },
+    {
+      id: 114,
+      name: '뭉클한',
+      type: 'IMPRESSION',
+      parentId: 101,
+      parentName: '긍정',
+      level: 2,
+      sortOrder: 4,
+      isSelectable: true,
+    },
+    {
+      id: 115,
+      name: '후련한',
+      type: 'IMPRESSION',
+      parentId: 101,
+      parentName: '긍정',
+      level: 2,
+      sortOrder: 5,
+      isSelectable: true,
+    },
+    {
+      id: 116,
+      name: '벅찬',
+      type: 'IMPRESSION',
+      parentId: 101,
+      parentName: '긍정',
+      level: 2,
+      sortOrder: 6,
+      isSelectable: true,
+    },
+    {
+      id: 117,
+      name: '안도한',
+      type: 'IMPRESSION',
+      parentId: 101,
+      parentName: '긍정',
+      level: 2,
+      sortOrder: 7,
+      isSelectable: true,
+    },
+    {
+      id: 118,
+      name: '희망이 생긴',
+      type: 'IMPRESSION',
+      parentId: 101,
+      parentName: '긍정',
+      level: 2,
+      sortOrder: 8,
+      isSelectable: true,
+    },
+    {
+      id: 119,
+      name: '설레는',
+      type: 'IMPRESSION',
+      parentId: 101,
+      parentName: '긍정',
+      level: 2,
+      sortOrder: 9,
+      isSelectable: true,
+    },
+    {
+      id: 120,
+      name: '흥미로운',
+      type: 'IMPRESSION',
+      parentId: 101,
+      parentName: '긍정',
+      level: 2,
+      sortOrder: 10,
+      isSelectable: true,
+    },
+    {
+      id: 121,
+      name: '빠져든',
+      type: 'IMPRESSION',
+      parentId: 101,
+      parentName: '긍정',
+      level: 2,
+      sortOrder: 11,
+      isSelectable: true,
+    },
+
+    // ============================================================
+    // 감상 키워드 (level 2, parentId: 102)
+    // ============================================================
+    {
+      id: 122,
+      name: '여운이 남는',
+      type: 'IMPRESSION',
+      parentId: 102,
+      parentName: '감상',
+      level: 2,
+      sortOrder: 1,
+      isSelectable: true,
+    },
+    {
+      id: 123,
+      name: '먹먹한',
+      type: 'IMPRESSION',
+      parentId: 102,
+      parentName: '감상',
+      level: 2,
+      sortOrder: 2,
+      isSelectable: true,
+    },
+    {
+      id: 124,
+      name: '울컥한',
+      type: 'IMPRESSION',
+      parentId: 102,
+      parentName: '감상',
+      level: 2,
+      sortOrder: 3,
+      isSelectable: true,
+    },
+    {
+      id: 125,
+      name: '찡한',
+      type: 'IMPRESSION',
+      parentId: 102,
+      parentName: '감상',
+      level: 2,
+      sortOrder: 4,
+      isSelectable: true,
+    },
+    {
+      id: 126,
+      name: '그리운',
+      type: 'IMPRESSION',
+      parentId: 102,
+      parentName: '감상',
+      level: 2,
+      sortOrder: 5,
+      isSelectable: true,
+    },
+    {
+      id: 127,
+      name: '익숙한',
+      type: 'IMPRESSION',
+      parentId: 102,
+      parentName: '감상',
+      level: 2,
+      sortOrder: 6,
+      isSelectable: true,
+    },
+    {
+      id: 128,
+      name: '이해가 되는',
+      type: 'IMPRESSION',
+      parentId: 102,
+      parentName: '감상',
+      level: 2,
+      sortOrder: 7,
+      isSelectable: true,
+    },
+    {
+      id: 129,
+      name: '의문이 남는',
+      type: 'IMPRESSION',
+      parentId: 102,
+      parentName: '감상',
+      level: 2,
+      sortOrder: 8,
+      isSelectable: true,
+    },
+
+    // ============================================================
+    // 부정 키워드 (level 2, parentId: 103)
+    // ============================================================
+    {
+      id: 131,
+      name: '지루한',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 1,
+      isSelectable: true,
+    },
+    {
+      id: 132,
+      name: '씁쓸한',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 2,
+      isSelectable: true,
+    },
+    {
+      id: 133,
+      name: '허무한',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 3,
+      isSelectable: true,
+    },
+    {
+      id: 134,
+      name: '찝찝한',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 4,
+      isSelectable: true,
+    },
+    {
+      id: 135,
+      name: '공허한',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 5,
+      isSelectable: true,
+    },
+    {
+      id: 136,
+      name: '서글픈',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 6,
+      isSelectable: true,
+    },
+    {
+      id: 137,
+      name: '분노가 이는',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 7,
+      isSelectable: true,
+    },
+    {
+      id: 138,
+      name: '복잡한',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 8,
+      isSelectable: true,
+    },
+    {
+      id: 139,
+      name: '허탈한',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 9,
+      isSelectable: true,
+    },
+    {
+      id: 140,
+      name: '불안한',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 10,
+      isSelectable: true,
+    },
+    {
+      id: 141,
+      name: '괴로운',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 11,
+      isSelectable: true,
+    },
+    {
+      id: 142,
+      name: '안타까운',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 12,
+      isSelectable: true,
+    },
+    {
+      id: 143,
+      name: '답답한',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 13,
+      isSelectable: true,
+    },
+    {
+      id: 144,
+      name: '슬픈',
+      type: 'IMPRESSION',
+      parentId: 103,
+      parentName: '부정',
+      level: 2,
+      sortOrder: 14,
+      isSelectable: true,
+    },
+  ],
+}
+
+// ============================================================
+// Mock Helper Functions
+// ============================================================
+
+/** 목데이터 응답 지연 시뮬레이션 (ms) */
+const MOCK_DELAY = 500
+
+const delay = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms))
+
+/**
+ * 키워드 목록 목데이터 반환
+ */
+export const getMockKeywords = async (): Promise<GetKeywordsResponse> => {
+  await delay(MOCK_DELAY)
+  return mockKeywordsResponse
+}
